mcp-server/requirements.txt
fastmcp==2.14.1
python-dotenv==1.2.1
google-cloud-bigquery==3.25.0
#google-genai>=1.50.0
